[LUA] Create job_utils/thief, migrate JA functionality, and add Larceny functionality
