+++ scripts/globals/abilities/collaborator.lua
@@ -5,22 +5,16 @@
 -- Recast Time: 1:00
 -- Duration: Instant
 -----------------------------------
-require("scripts/globals/settings")
-require("scripts/globals/status")
-require("scripts/globals/msg")
+require("scripts/globals/job_utils/thief")
 -----------------------------------
 local ability_object = {}
 
 ability_object.onAbilityCheck = function(player, target, ability)
-    if (target == nil or target:getID() == player:getID() or not target:isPC()) then
-        return xi.msg.basic.CANNOT_ON_THAT_TARG, 0
-    else
-        return 0, 0
-    end
+    return xi.job_utils.thief.checkCollaborator(player, target, ability)
 end
 
 ability_object.onUseAbility = function(player, target, ability)
-    target:transferEnmity(player, 25 + player:getMod(xi.mod.ACC_COLLAB_EFFECT), 20.6)
+    xi.job_utils.thief.useCollaborator(player, target, ability)
 end
 
 return ability_object
