+++ scripts/globals/abilities/feint.lua
@@ -5,8 +5,7 @@
 -- Recast Time: 2:00 minutes
 -- Duration: 1:00 minutes
 -----------------------------------
-require("scripts/globals/settings")
-require("scripts/globals/status")
+require("scripts/globals/job_utils/thief")
 -----------------------------------
 local ability_object = {}
 
@@ -15,8 +14,7 @@ ability_object.onAbilityCheck = function(player, target, ability)
 end
 
 ability_object.onUseAbility = function(player, target, ability)
-    local augment = player:getMod(xi.mod.AUGMENTS_FEINT) * player:getMerit(xi.merit.FEINT) / 25 -- Divide by the merit value (feint is 25) to get the number of merit points
-    player:addStatusEffect(xi.effect.FEINT, 150 + augment, 0, 60) -- -150 Evasion base
+    xi.job_utils.thief.useFeint(player, target, ability)
 end
 
 return ability_object
