+++ scripts/globals/abilities/flee.lua
@@ -5,7 +5,7 @@
 -- Recast Time: 5:00
 -- Duration: 0:30
 -----------------------------------
-require("scripts/globals/status")
+require("scripts/globals/job_utils/thief")
 -----------------------------------
 local ability_object = {}
 
@@ -14,8 +14,7 @@ ability_object.onAbilityCheck = function(player, target, ability)
 end
 
 ability_object.onUseAbility = function(player, target, ability)
-    local duration = 30 + player:getMod(xi.mod.FLEE_DURATION)
-    player:addStatusEffect(xi.effect.FLEE, 100, 0, duration)
+    xi.job_utils.thief.useFlee(player, target, ability)
 end
 
 return ability_object
