+++ scripts/globals/abilities/hide.lua
@@ -5,8 +5,7 @@
 -- Recast Time: 5:00
 -- Duration: Random
 -----------------------------------
-require("scripts/globals/settings")
-require("scripts/globals/status")
+require("scripts/globals/job_utils/thief")
 -----------------------------------
 local ability_object = {}
 
@@ -15,9 +14,7 @@ ability_object.onAbilityCheck = function(player, target, ability)
 end
 
 ability_object.onUseAbility = function(player, target, ability)
-    local duration = math.random(30, 300)
-    duration = duration * (1 + player:getMod(xi.mod.HIDE_DURATION)/100)
-    player:addStatusEffect(xi.effect.HIDE, 1, 0, math.floor(duration * xi.settings.main.SNEAK_INVIS_DURATION_MULTIPLIER))
+    xi.job_utils.thief.useHide(player, target, ability)
 end
 
 return ability_object
