+++ scripts/globals/abilities/larceny.lua
@@ -4,18 +4,16 @@
 -- Obtained: THF Level 96
 -- Recast Time: 01:00:00
 -----------------------------------
-require("scripts/globals/settings")
-require("scripts/globals/status")
+require("scripts/globals/job_utils/thief")
 -----------------------------------
 local ability_object = {}
 
 ability_object.onAbilityCheck = function(player, target, ability)
-    ability:setRecast(ability:getRecast() - player:getMod(xi.mod.ONE_HOUR_RECAST))
-    return 0, 0
+    return xi.job_utils.thief.checkLarceny(player, target, ability)
 end
 
-ability_object.onUseAbility = function(player, target, ability)
-    -- player:addStatusEffect(xi.effect.LARCENY, 6, 0, 5) -- TODO: implement xi.effect.LARCENY
+ability_object.onUseAbility = function(player, target, ability, action)
+    return xi.job_utils.thief.useLarceny(player, target, ability, action)
 end
 
 return ability_object
